Refactoring after review
